Move web3 functions into utils/web3 instead of generic utils file
