+++ src/commands/create/contract.ts
@@ -8,7 +8,7 @@ import {networks} from '@holographxyz/networks'
 
 import {BytecodeType, bytecodes} from '../../utils/bytecodes'
 import {ensureConfigFileIsValid} from '../../utils/config'
-import {web3, zeroAddress, remove0x, sha3, dropEventsEnabled} from '../../utils/utils'
+import {web3, zeroAddress, remove0x, sha3, dropEventsEnabled} from '../../utils/web3'
 import {NetworkMonitor} from '../../utils/network-monitor'
 import {
   ContractDeployment,
